+++ src/Message/DirectPostAbstractRequest.php
@@ -36,4 +36,26 @@ abstract class DirectPostAbstractRequest extends AbstractRequest
 
         return sha1($hash);
     }
+
+    /**
+     * @return mixed
+     */
+    public function getBaseData()
+    {
+        $data = array();
+
+        $data['EPS_MERCHANT'] = $this->getMerchantId();
+        $data['EPS_TXNTYPE'] = $this->txnType;
+        $data['EPS_IP'] = $this->getClientIp();
+        $data['EPS_AMOUNT'] = $this->getAmount();
+        $data['EPS_REFERENCEID'] = $this->getTransactionId();
+        $data['EPS_TIMESTAMP'] = gmdate('YmdHis');
+        $data['EPS_FINGERPRINT'] = $this->generateFingerprint($data);
+        $data['EPS_RESULTURL'] = $this->getReturnUrl();
+        $data['EPS_CALLBACKURL'] = $this->getNotifyUrl() ?: $this->getReturnUrl();
+        $data['EPS_REDIRECT'] = 'TRUE';
+        $data['EPS_CURRENCY'] = $this->getCurrency();
+
+        return $data;
+    }
 }
