+++ src/Message/DirectPostAuthorizeRequest.php
@@ -15,28 +15,6 @@ class DirectPostAuthorizeRequest extends DirectPostAbstractRequest
     /**
      * @return mixed
      */
-    public function getBaseData()
-    {
-        $data = array();
-
-        $data['EPS_MERCHANT'] = $this->getMerchantId();
-        $data['EPS_TXNTYPE'] = $this->txnType;
-        $data['EPS_IP'] = $this->getClientIp();
-        $data['EPS_AMOUNT'] = $this->getAmount();
-        $data['EPS_REFERENCEID'] = $this->getTransactionId();
-        $data['EPS_TIMESTAMP'] = gmdate('YmdHis');
-        $data['EPS_FINGERPRINT'] = $this->generateFingerprint($data);
-        $data['EPS_RESULTURL'] = $this->getReturnUrl();
-        $data['EPS_CALLBACKURL'] = $this->getNotifyUrl() ?: $this->getReturnUrl();
-        $data['EPS_REDIRECT'] = 'TRUE';
-        $data['EPS_CURRENCY'] = $this->getCurrency();
-
-        return $data;
-    }
-
-    /**
-     * @return mixed
-     */
     public function getData()
     {
         $this->validate('amount', 'returnUrl', 'card');
